`coredns`: run `setcap` so that the binary can bind to low ports as `nonroot` (#235)
The default `sysctl` profile on Kubernetes disallows nonroot users from binding to low ports, but with `setcap` we can bind to a low port without needing to configure the "safe" `sysctl` in kubernetes.

Without this, our `coredns` image fails to start (without additional configuration), but with this I was able to make it run without any additional configuration.

Once this lands, I will expand the downstream image test to test things with the stricter k8s-compatible sysctl setting.